+++ app/views/admin/cuentas/index.blade.php
@@ -10,7 +10,7 @@
 
 @section('breadcrumb')
     @parent
-    <li class="active"><a href="{{ url('admin/cuentas')  }}">Cuenta</a></li>
+    <li class="active"><a href="{{ url('admin/cuentas')  }}">Cuentas</a></li>
 @endsection
 
 @section('content')
@@ -41,22 +41,37 @@
             </thead>
             <tbody>
             @foreach ($cuentas as $cliente)
-                <tr>
-                    <td>{{ $cliente->rut_cliente }}</td>
-                    <td>{{ $cliente->nombre_cliente }}</td>
-                    <td>{{ $cliente->fono_fijo_cliente }}</td>
-                    <td>{{ $cliente->correo_cliente }}</td>
-                    <td class="pull-right">
-                        {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin.cuentas.destroy', $cliente->id_cliente))) }}
-                        {{ Form::submit('Eliminar', array('class' => 'btn btn-danger')) }}
-                        {{ Form::close() }}
-                        {{ link_to_route('admin.cuentas.edit', 'Editar', array($cliente->id_cliente), array('class' => 'btn btn-info')) }}
-                        {{ link_to_action('CuentasController@resumen', 'Resumen', array($cliente->id_cliente), array('class' => 'btn btn-link')) }}
-                    </td>
-                </tr>
+                @if($cliente->id_cliente != 1)
+                    <tr>
+                        <td>{{ $cliente->rut_cliente }}</td>
+                        <td>{{ $cliente->nombre_cliente }}</td>
+                        <td>{{ $cliente->fono_fijo_cliente }}</td>
+                        <td>{{ $cliente->correo_cliente }}</td>
+                        <td class="pull-right">
+                            {{ Form::open(array('style' => 'display: inline-block;', 'method' => 'DELETE', 'route' => array('admin.cuentas.destroy', $cliente->id_cliente))) }}
+                            {{ Form::submit('Eliminar', array('class' => 'btn btn-danger')) }}
+                            {{ Form::close() }}
+                            {{ link_to_route('admin.cuentas.edit', 'Editar', array($cliente->id_cliente), array('class' => 'btn btn-info')) }}
+                            {{ link_to_action('CuentasController@resumen', 'Resumen', array($cliente->id_cliente), array('class' => 'btn btn-link')) }}
+                        </td>
+                    </tr>
+                @else
+                    <tr>
+                        <td></td>
+                        <td>{{ $cliente->nombre_cliente }}</td>
+                        <td>{{ $cliente->fono_fijo_cliente }}</td>
+                        <td>{{ $cliente->correo_cliente }}</td>
+                        <td class="pull-right">
+                            {{ link_to_route('admin.cuentas.edit', 'Editar', array($cliente->id_cliente), array('class' => 'btn btn-info')) }}
+                            {{ link_to_action('CuentasController@resumen', 'Resumen', array($cliente->id_cliente), array('class' => 'btn btn-link')) }}
+                        </td>
+                    </tr>
+                @endif
             @endforeach
             </tbody>
         </table>
+
+        {{ $cuentas->links() }}
     @else
         No se han encontrado cuentas.
     @endif
